Setting 'apacite' as default bibliography style
With apacite, the format for citations in the slides has the
form (author, year), which makes it easy for viewer to remember
the citation and look it up afterwards.

With other styles (e.g. [1]) a viewer has to remember a number
or similar and then look it up later in the bibliography - this
way a viewer can just note down the author and name directly and
search for the reference later (e.g. through a search engine).